feat(link): Replace mDNS with a functional configuration interface
Summary: This diff removes the existing mDNS discovery solution for bridges and replaces it with a functional localhost web interface in link to allow the end user to configure the connection to a core. This diff will also provide the base configuration for automatic link builds and artifact distribution via Harbourmaster

Test Plan: Go through a conventional workflow from both a blank slate and a preconfigured state to ensure no edge cases appear from a half-configured or orphan situation. This needs to be tested both in a development environment and as a built distributable package external to the development environment to ensure all dependencies are properly bundled and packaged.

Reviewers: monoxane

Reviewed By: monoxane

Differential Revision: https://phabricator.boreal.systems/D13